Added custom hooks in the Create Flashcard Page
