+++ src/pages/CreateFlashcards/hooks/useDescription.tsx
@@ -1,0 +1,9 @@
+import { useState } from "react";
+
+export const useDescription = () => {
+  const [description, setDescription] = useState("");
+  const handleDescriptionChange = (e: React.ChangeEvent<HTMLTextAreaElement>) =>
+    setDescription(e.target.value);
+
+  return { description, setDescription, handleDescriptionChange };
+};
