+++ src/pages/CreateFlashcards/hooks/useTitle.tsx
@@ -1,0 +1,9 @@
+import { useState } from "react";
+
+export const useTitle = () => {
+  const [title, setTitle] = useState("");
+  const handleTitleChange = (e: React.ChangeEvent<HTMLInputElement>) =>
+    setTitle(e.target.value);
+
+  return { title, setTitle, handleTitleChange };
+};
